backend/requirements.txt
django==2.2.16
djangorestframework==3.12.4
Pillow==9.0.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
reportlab==3.6.12
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1


